feat: implement menu building functionality and update menu file path
