+++ apps/react-router/scripts/build-index.ts
@@ -1,3 +1,5 @@
 import { buildIndex } from '@remix-docs-ja/scripts/build-index'
+import { buildMenus } from '@remix-docs-ja/scripts/build-menu'
 
 await buildIndex('react-router-v7')
+await buildMenus()
